Extract `Projects::Identifier` concern from `Project` model
Move all identifier-related logic out of the Project model into a
dedicated Projects::Identifier concern. This includes:

- Constants (IDENTIFIER_MAX_LENGTH, SEMANTIC_IDENTIFIER_MAX_LENGTH,
  RESERVED_IDENTIFIERS)
- FriendlyId configuration and slug management
- acts_as_url configuration for identifier generation
- All identifier format validators (legacy and alphanumeric)
- Historical identifier reservation check
- suggest_identifier class method
- validation_context override for :saving_custom_fields

The unset_slug_if_invalid override is moved inside the `included`
block to correctly override FriendlyId::Slugged in the MRO.

Corresponding tests are moved from project_spec.rb to a new
projects/identifier_spec.rb.

No behavior changes — pure extraction and reorganization.
