+++ spec/models/project_spec.rb
@@ -557,123 +557,6 @@ RSpec.describe Project do
     end
   end
 
-  describe "url identifier" do
-    let(:reserved) do
-      Rails.application.routes.routes
-        .map { |route| route.path.spec.to_s }
-        .filter_map { |path| path[%r{^/projects/(\w+)\(\.:format\)$}, 1] }
-        .uniq
-    end
-
-    it "is set from name" do
-      project = described_class.new(name: "foo")
-
-      project.validate
-
-      expect(project.identifier).to eq("foo")
-    end
-
-    it "is not allowed to clash with projects routing" do
-      expect(reserved).not_to be_empty
-
-      reserved.each do |word|
-        project = described_class.new(name: word)
-
-        project.validate
-
-        expect(project.identifier).not_to eq(word)
-      end
-    end
-
-    it "is not allowed to clash with another project" do
-      create(:project, identifier: "existing")
-
-      project = build(:project, identifier: "existing")
-      expect(project).not_to be_valid
-      expect(project.errors[:identifier]).to include("has already been taken.")
-    end
-
-    it "is not allowed to clash with a former identifier of another project" do
-      other_project = create(:project, identifier: "former-id")
-      other_project.update!(identifier: "new-id")
-
-      project = build(:project, identifier: "former-id")
-      expect(project).not_to be_valid
-      expect(project.errors[:identifier]).to include("has already been taken.")
-    end
-
-    it "is allowed to be the same as its own former identifier" do
-      project = create(:project, identifier: "old-id")
-      project.update!(identifier: "new-id")
-
-      project.identifier = "old-id"
-      expect(project).to be_valid
-    end
-
-    # The acts_as_url plugin defines validation callbacks on :create and it is not automatically
-    # called when calling a custom context. However we need the acts_as_url callback to set the
-    # identifier when the validations are called with the :saving_custom_fields context.
-    context "when validating with :saving_custom_fields context" do
-      it "is set from name" do
-        project = described_class.new(name: "foo")
-
-        project.validate(:saving_custom_fields)
-
-        expect(project.identifier).to eq("foo")
-      end
-
-      it "is not allowed to clash with projects routing" do
-        expect(reserved).not_to be_empty
-
-        reserved.each do |word|
-          project = described_class.new(name: word)
-
-          project.validate(:saving_custom_fields)
-
-          expect(project.identifier).not_to eq(word)
-        end
-      end
-    end
-
-    context "with history" do
-      let!(:project) { create(:project, identifier: "sc") }
-
-      it "records the old identifier in friendly_id_slugs when identifier changes" do
-        project.update!(identifier: "scp")
-        expect(FriendlyId::Slug.where(sluggable: project).pluck(:slug)).to include("sc")
-      end
-
-      it "can still find the project via its old identifier" do
-        project.update!(identifier: "scp")
-        expect(described_class.friendly.find("sc")).to eq(project)
-      end
-
-      it "returns the project with its current identifier when found via old identifier" do
-        project.update!(identifier: "scp")
-        found = described_class.friendly.find("sc")
-        expect(found.identifier).to eq("scp")
-      end
-
-      it "locks old identifier to the original project (not reusable by others)" do
-        project.update!(identifier: "scp")
-        slug = FriendlyId::Slug.find_by(slug: "sc")
-        expect(slug.sluggable_id).to eq(project.id)
-      end
-
-      it "allows the project to revert to a previously used identifier" do
-        project.update!(identifier: "scp")
-        expect { project.update!(identifier: "sc") }.not_to raise_error
-        expect(project.identifier).to eq("sc")
-      end
-
-      it "is valid when reverting to own historical identifier" do
-        project.update!(identifier: "scp")
-        project.identifier = "sc"
-        expect(project).to be_valid
-      end
-    end
-  end
-
   describe "#allowed_parent_workspace_types" do
     {
       project: %i[portfolio program project],
@@ -724,21 +607,4 @@ RSpec.describe Project do
     end
   end
 
-  describe ".suggest_identifier" do
-    context "with alphanumeric identifiers", with_settings: { work_packages_identifier: "alphanumeric" } do
-      it "delegates to ProjectIdentifierSuggestionGenerator" do
-        allow(WorkPackages::IdentifierAutofix::ProjectIdentifierSuggestionGenerator)
-          .to receive(:suggest_identifier).with("My Project").and_return("MP")
-        expect(described_class.suggest_identifier("My Project")).to eq("MP")
-        expect(WorkPackages::IdentifierAutofix::ProjectIdentifierSuggestionGenerator)
-          .to have_received(:suggest_identifier).with("My Project")
-      end
-    end
-
-    context "with numeric (legacy) identifiers", with_settings: { work_packages_identifier: "numeric" } do
-      it "returns a slugified lowercase identifier" do
-        expect(described_class.suggest_identifier("My Cool Project")).to eq("my-cool-project")
-      end
-    end
-  end
 end
